add ergonomic functions for node iteration
Makes it so that we don't have to keep track
of the node ID ourselves.

I considered just passing a reference to the
MapNode, but it seems like it's often useful
for callers to know the ID of the node in
question (eg. for finding clockwise neighbours,
or for printing the ID)